Update
Currently basic chat for red/blue rooms seems to work. Private message to other user seems to go through but still need to figure out the logic to make it work better on the front-end.
Currently private message will change the room instantly to other room for the receiver and sender. Which is not cool user-experience.